+++ R/permute_group.R
@@ -70,9 +70,9 @@ permute.group <- function(permSet, density, resids,
   groups <- as.numeric(resids$Group)
   out <- foreach(i=seq_len(nrow(permSet)), .combine='rbind',
                  .export='assign_lobes') %dopar% {
-    corrs1 <- corr.matrix(resids[which(groups[permSet[i, ]] == 1), !'Group', with=F],
+    corrs1 <- corr.matrix(resids[which(groups[permSet[i, ]] == 1)],
                           density=density)
-    corrs2 <- corr.matrix(resids[which(groups[permSet[i, ]] == 2), !'Group', with=F],
+    corrs2 <- corr.matrix(resids[which(groups[permSet[i, ]] == 2)],
                           density=density)
     g1 <- graph_from_adjacency_matrix(corrs1$r.thresh, mode='undirected', diag=F)
     g2 <- graph_from_adjacency_matrix(corrs2$r.thresh, mode='undirected', diag=F)