+++ NEWS.md
@@ -1,6 +1,9 @@
-# brainGraph 0.52.0
+# brainGraph 0.52.1
 
 ## Minor changes
+* `brainGraph_init` does a better job of dealing with subcortical gray matter data
+* `boot_global` and `permute.group` updated to work with new version of `corr.matrix`
+* `get.resid` now returns *data.table* of residuals with a *Study.ID* column
 * `corr.matrix`: minor syntax change
 * `get.resid`: fixed minor bug when `use.mean=FALSE` but *covars* has columns
     *mean.lh* and/or *mean.rh*
@@ -18,3 +21,10 @@
 * `permute.group` no longer takes `atlas.dt` as an argument
 * `vertex_attr_dt` is now essentially a wrapper for `igraph`'s function
     `as_data_frame`
+
+
+# brainGraph 0.48.0
+
+2015-12-08
+
+Initial CRAN acceptance